common/requirements.txt
mypy==0.790
python-dateutil==2.8.1
flake8==3.8.4
psycopg2-binary==2.8.6
pylint==2.6.0
pytest==6.1.2
SQLAlchemy==1.3.20
sqlalchemy-stubs==0.3
typedload==2.3
yoyo-migrations==7.2.0
python-dotenv==0.15.0
minio==6.0.0
-e .